Only start spring if toValue has changed
Summary:
I have a hypothesis that the only time the spring should automatically start is if `toValue` changes.  If the spring is already animating when the config is updated, it will continue to animate with new parameters.  If it is already at rest when a setting or an initial value is changed, the spring should probably stay put.

Test Plan:

Reviewers:

Subscribers: