Moved minified bundles to .min.js
Summary:
Presently, `yarn run dist` generates minified bundles.  This commit makes the default bundles unminified, and  adds a `.min.js` variant to each.

This change helps with downstream bundling.  Running Rollup with the `resolve` plugin will not inline the minified Wobble, but it will inline the unminified Wobble.  I also presume that it's easier for a bundler to handle unminified source, so it can make its optimizations with the full context of the bundle and avoid multiple polyfills/etc.

Reviewers: skevy

Reviewed By: skevy

Differential Revision: http://codereview.cc/D3500